@@ -1,5 +1,5 @@
 import React from "react";
-import Link from "next/link";
+import { Link } from "react-router-dom";
 
 export default function BlogPreviewItem(props) {
   const { blog } = props;
@@ -8,7 +8,7 @@ export default function BlogPreviewItem(props) {
 
   return (
     <Link
-      href={`/blog/${blog.id}`}
+      to={`/blog/${blog.id}`}
       className="block border-l border-l-4 border-gray-300 hover:border-yellow-400 mb-2 p-2 pointer transition duration-500 ease-in-out"
     >
       <div className="block md:flex text-gray-500 mb-1">
